// src/app/maps/interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
    type:        string;
    query:       string[];
    features:    Feature[];
    attribution: string;
}

export interface Feature {
    id:            string;
    type:          string;
    place_type:    string[];
    relevance:     number;
    properties:    Properties;
    text_en:       string;
    place_name_en: string;
    text:          string;
    place_name:    string;
    center:        number[];
    geometry:      Geometry;
    context:       Context[];
}

export interface Context {
    id:           string;
    mapbox_id:    string;
    text_en:      Text;
    text:         Text;
    wikidata?:    Wikidata;
    language_en?: Language;
    language?:    Language;
    short_code?:  ShortCode;
}

export enum Language {
    En = "en",
}

export enum ShortCode {
    Ar = "ar",
    ArS = "AR-S",
}

export enum Text {
    Argentina = "Argentina",
    S3000 = "S3000",
    S3002 = "S3002",
    SantaFe = "Santa Fe",
}

export enum Wikidata {
    Q414 = "Q414",
    Q44244 = "Q44244",
    Q44823 = "Q44823",
}

export interface Geometry {
    coordinates: number[];
    type:        string;
}

export interface Properties {
    foursquare: string;
    landmark:   boolean;
    address?:   string;
    category:   string;
    maki?:      string;
}